AddInt.java

+++ AddInt.java
@@ -2,8 +2,8 @@
 
 public class AddInt{
 	public static void main(String [] args){
-		int a = 12;
-		int b = 38;
+		double a = 12.0;
+		double b = 38.0;
 		double Avg = (a + b)/ 2.0;
 		System.out.println( Avg);
